requirements.txt
playwright~=1.28.0
PyYAML~=6.0